+++ ClimaTempo/shared/src/commonMain/kotlin/me/rafael/yokota/shared/viewmodel/Greeting.kt
@@ -4,7 +4,7 @@ import api_weather.WeatherSDK
 
 
 class Greeting {
-    suspend fun greeting(lat: Float, lon: Float): String {
+    suspend fun greeting(lat: Double, lon: Double): String {
         return WeatherSDK().getWeatherDetails(lat,lon).toString()
     }
 }
